// types.ts
export type FlightMode = 'C' | 'N' | 'S';